backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
google-genai>=0.8.0
